requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
starlette==0.38.6
pywebview==5.2; platform_system == "Windows"
pythonnet==3.0.5; platform_system == "Windows"
